Bump production Jenkins pod resources to match current instance size
The current OpenSearch Jenkins controller node is a c5.9xlarge (36c, 72GB).
with a 100GB disk. This bumps the Jenkins controller pod resource requests
to match that, less a bit to let it fit within a 9xlarge Kubernetes node
after the K8S node system-reserved resource overhead.

This also bumps the Java min/max heap size flags to use this higher memory,
but this is a guess and should probably be tuned later.

Signed-off-by: Robert Detjens <rdetjens@linuxfoundation.org>